requirements.txt
# LLM
httpx>=0.27.0

# STT
faster-whisper>=1.0.3

# TTS online (fallback)
edge-tts>=7.0.0

# Áudio
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26.0
pygame-ce>=2.5.0

# Câmera + tela
opencv-python>=4.9.0
mss>=9.0.1

# Controle de volume
pycaw>=20240210
comtypes>=1.4.1

# Controle de processos
psutil>=5.9.0

# Teclado e atalhos
keyboard>=0.13.5

# Clipboard
pyperclip>=1.8.2

# Brilho (opcional — pode falhar em monitores externos)
screen-brightness-control>=0.23.0

# Web search
googlesearch-python>=1.2.0
ddgs>=0.1.0
# duckduckgo-search>=6.2.0  # renomeado para ddgs
beautifulsoup4>=4.12.0

# Hardware monitoring
GPUtil>=1.4.0

# Notificações Windows
winotify>=1.1.0

# TTS SAPI (Windows offline — fallback)
pyttsx3>=2.90

# OCR — lê texto da tela
pytesseract>=0.3.10

# Spotify Web API
spotipy>=2.23.0

# Histórico SQLite (built-in, sem dependência extra)

# RAG — memória de longo prazo
chromadb>=0.5.0

# Interface Web
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Variáveis de ambiente (.env)
python-dotenv>=1.0.0

# Google Calendar (opcional)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Resumo de páginas
# youtube-transcript-api>=0.6.2  # opcional: resumir vídeos do YouTube
